move utility functions/constructs to utils/
